requirements.txt
boto3==1.28.63
Django==4.1.7
djangorestframework==3.14.0
pynamodb==5.5.0
python-dotenv==1.0.0
